requirements.txt

torch==2.0.0
torchvision==0.15.1
transformers==4.28.1
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.2
Flask==2.2.5
Werkzeug==2.2.3
